backend/requirements.txt
Flask==2.0.3
uWSGI==2.0.20
Flask-SQLAlchemy==2.5.1
Flask-Marshmallow==0.14.0
Flask-Restless-NG==2.2.9
psycopg2==2.9.3
marshmallow-sqlalchemy==0.27.0
flask_cors==3.0.10